- Add unit tests for GithubUser, GithubOrganization, GithubRepository - Implements equals() and hashcode() for GithubRepository, GithubPullRequest, GithubOrganization, GithubUser

+++ plugins/com.aptana.git.core/src/com/aptana/git/core/github/IGithubOrganization.java
@@ -17,7 +17,7 @@ import org.eclipse.core.runtime.CoreException;
 public interface IGithubOrganization
 {
 
-	public int getID();
+	public long getID();
 
 	/**
 	 * Name of the org.
